feat: event

+++ src/graph/event/delete.ts
@@ -1,4 +1,5 @@
 import Graph from "..";
+import { EVENT_UNLINK } from "../event";
 import KeyValueBox from "../keyvalueBox";
 import LinkLine from "../LinkLine";
 import ObjectBox from "../ObjectBox";
@@ -8,5 +9,9 @@ export default function deleteItem(
   item: LinkLine | ObjectBox | KeyValueBox
 ) {
   if (!item) return;
-  item.delete();
+  if (item instanceof LinkLine) {
+    graph.emit(EVENT_UNLINK, { line: item });
+  } else {
+    item.delete();
+  }
 }
